APIリクエストをモデルではなく、コントローラーで行うように変更した
